refactor: auth 패키지 구조화
