+++ src/main/java/com/stayforyou/auth/dto/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package com.stayforyou.auth;
+package com.stayforyou.auth.dto;
 
 import java.util.Collection;
 import java.util.List;
