+++ src/main/java/com/stayforyou/auth/filter/JwtFilter.java
@@ -1,0 +1,4 @@
+package com.stayforyou.auth.filter;
+
+public class JwtFilter {
+}
