+++ src/test/java/com/stayforyou/auth/JwtUtilTest.java
@@ -2,6 +2,7 @@ package com.stayforyou.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.stayforyou.auth.util.JwtUtil;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
